refactor: user-profile
