+++ src/app/ranking/ranking-prefecture/page.tsx
@@ -7,9 +7,9 @@ import { PrefectureSelect } from "@/features/ranking/components/prefecture-selec
 import { RankingPrefecture } from "@/features/ranking/components/ranking-prefecture";
 import { RankingTabs } from "@/features/ranking/components/ranking-tabs";
 import { getUserPrefecturesRanking } from "@/features/ranking/services/get-prefectures-ranking";
+import { getMyProfile } from "@/features/user-profile/services/profile";
 import { PREFECTURES } from "@/lib/address";
 import { getCurrentSeasonId } from "@/lib/services/seasons";
-import { getMyProfile } from "@/lib/services/users";
 import { createClient } from "@/lib/supabase/client";
 
 interface PageProps {
