+++ src/components/hero.tsx
@@ -1,6 +1,6 @@
 import { OnboardingButton } from "@/features/onboarding/components/onboarding-button";
 import Levels from "@/features/user-level/components/levels";
-import { getUser } from "@/lib/services/users";
+import { getUser } from "@/features/user-profile/services/profile";
 import Image from "next/image";
 import Link from "next/link";
 import { Button } from "./ui/button";
