+++ src/features/user-profile/components/user-avatar.tsx
